BoxedUint: move `From` impls to module
More consistent with `Uint`
